pytest: Split out REDCap tests
